Add basic varnishkafka rsyslog config to the varnishkafka module.
The varnishkafka_rsyslog.conf is stored at the moment in the main
puppet repository and referenced/created by the role cache::kafka.
This patch adds basic rsyslog configuration for the varnishkafka
daemon without relying on any other entity (like roles) that will
have only to deal with varnishkafka instances.
Logrotate configuration for the basic daemon is contained in this
module too.

Bug: T129344
Change-Id: I91a7a10c33ac3c2a64fba0a79f30bef3ead92730