varnishkafka.py: Set owner/group/mode
Not specifying the mode attribute on a file resource means the mode
attribute on the resulting file will be what is on the puppetmaster.
That, aside from being incomplete and wrong, allows for the possibility
to have the file mode changed depending on what the file is on the
backend puppetmaster leading to harmless but spurious and useless
logging. Fix that by specifying 0444

Change-Id: I6f8b91c1b98bef4ac958231621ed3c11691b3927